Update README.md
The objective of this project is to design and develop a prototype of an industrial robot (robotic arm) with four degrees of freedom, controlled by a radio-controlled programmable logic controller. It also allows for monitoring data processing and the radio channel through which the two components of the project communicate with each other. The robot reproduces the user's hand movements using sensors. The main objective is to develop an initial prototype that not only demonstrates key features but also serves as a basis for the subsequent development of a more complex and advanced system, suitable for specific practical tasks and offering potential for practical application in the relevant field. This prototype allows for the validation of the main subsystems, such as servomotor control via a microcontroller. The project consists of two main parts: transmission and reception. The transmitter records hand movements, processes the collected data, and transmits commands via radio to the transmitter and then to the receiver, to which the mechanical robotic hand is connected, executing the corresponding movements. Each hand gesture 5 corresponds to a code generated by the transmitter, transmitted by radio to the transmitter and executed by the latter.